requirements.txt
discord.py==2.4.0
aiosqlite==0.19.0
python-dotenv==1.0.0
flask==3.0.0
aiofiles==23.2.1
PyNaCl==1.5.0
audioop-lts==0.2.1
aiohttp==3.9.5
